Ensure we flag newer locally moved children for reupload.
This likely explains why we're seeing more `parentid` mismatches for
the new bookmarks engine. Previously, when we preferred a local move
over a remote move, we flagged the *parent* for reupload, but not the
*child*. However, older Desktops and Android use the child's `parentid`
to determine the correct parent, not the parent's `children`, and iOS
expects the two to match.

The old behavior accidentally did the right thing on Desktop for local
values, since we didn't reset the Sync change counter (`needs_merge`
flag in Dogear). However, we did reset the counter for *remote* values
with a new structure, meaning those children weren't uploaded to the
server...even though they were flagged correctly before merging.

This commit also simplifies Sync change tracking for uploads. Every
merged node wiith a new structure (`LocalWithNewStructure` or
`RemoteWithNewStructure`) should have `syncChangeCounter = 1`, and
all others should have `syncChangeCounter = 0`.