Replace `log` with a simple local logger.
The `log` crate uses a global logger that can only be registered once,
at startup. Unfortunately, this won't work for Desktop, where we want
to integrate with Sync's existing logging setup. We also don't want to
impact other crates in `gkrust` that use `log`.

This commit replaces `log` with a quick-and-dirty implementation on
`Driver`, and some macros that take care of calling `format_args!()`.
We can extend these later to pass `file!()` and `line!()`, if we
want.

This commit also changes `Merger` to hold a reference to `Driver`.
On Desktop, the driver accumulates log messages on the storage
thread, and accesses them on the main thread.