Replace `Stats` with a struct returned from `Store::merge`.
Since we already tracked most of our counts in `StructureCounts`, this
commit adds `merged_nodes` and `merged_deletions`. Bonus: counting
merged nodes this way lets us preallocate the space we need to hold all
the descendants when we recurse down the merged root.

This also adds a `MergedTimings` struct, and a macro for recording
timings with operation names.