Add an `UploadReason` for merged descendants.
We'll use this on Desktop to track sync loops, where two or more
clients reupload the same record on every sync, even though the
user hasn't changed it locally.

The idea is: on every sync, we'll store the GUIDs of all uploaded items
in a separate table, along with the timestamp and reason. If we see a
GUID uploaded with `NewStructure` for 5-ish consecutive syncs, we'll
emit telemetry.