Factor out resolving possible parents into `PossibleParents`.
The logic for breaking the tie between conflicting parents is hard to
digest when it's inlined into `resolve`. Instead of using `min_by()`,
we can iterate over a sequence of `PossibleParent`s that implement
the logic as part of `Ord::cmp`.

This commit also adds more comments that explain how we pick the
parent.