Factor out parent resolution into `ResolveParent`.
This builds on the last commit. The algorithm for resolving parents
is hard to follow when it's inlined into `Builder::resolve`, isn't
as efficient as it could be for roots, isn't well-commented, and
tolerates unreachable cases.

This commit factors that logic out into a separate `ResolveParent`
type. It simplifies how we handle complete and partial parents,
unrolls the nested iterator chains in `Builder::into_tree`,
tightens up some cases, and adds more docs.