Record and report problems with tree structure.
This commit records problems that Dogear notices when building a
tree, and outputs a summary when the tree is printed. It's useful for
logging, so we can see why the structure diverged.

We can also use it to emit validation events, like the Desktop Sync
ping, if we decide to in the future.