Remove `Store::fetch_new_{local, remote}_contents`.
In Rust Places and `bookmark_sync` on Desktop, fetching the tree and
contents runs two table scans: one to build the structure, and one to
fetch content info for deduping. Since the latter can't use an index,
it's more efficient to fetch content info during the first scan to
build the tree.

This commit adds a builder API to set content info, and stores it with
the item instead of a separate map.