Revert storing extra tombstones in `delete_{locally, remotely}`.
Instead of panicking if a GUID in `delete_locally` or `delete_remotely`
doesn't exist on both sides, we now clean up local tombstones, and
flag the remote tombstone as merged.

With this change, we no longer need to filter out local and remote
tombstones for items that don't exist on the other side. That means
we can revert `Tree::guids` to return all GUIDs, including deletions,
and have `MergedRoot::{local, remote}_deletions` return a more accurate
set of tombstones for logging.